pip_requirements.txt
django-autoslug
pycountry
